+++ GPTifier/CMakeLists.txt
@@ -80,7 +80,6 @@ add_executable(
   "gpt"
   "src/cli.cpp"
   "src/command_chats.cpp"
-  "src/command_costs.cpp"
   "src/command_files.cpp"
   "src/command_fine_tune.cpp"
   "src/command_models.cpp"
@@ -88,6 +87,7 @@ add_executable(
   "src/configs.cpp"
   "src/datadir.cpp"
   "src/help_messages.cpp"
+  "src/interface/command_costs.cpp"
   "src/interface/command_embed.cpp"
   "src/interface/command_run.cpp"
   "src/interface/command_short.cpp"
