+++ GPTifier/include/interface/command_costs.hpp
